feat: implement basic auth

+++ src/controllers/document_controller.py
@@ -1,7 +1,8 @@
 import traceback
 from typing import Annotated
-from fastapi import APIRouter, Depends, Form, HTTPException, File, UploadFile
+from fastapi import APIRouter, Depends, Form, HTTPException, File, Header, UploadFile
 
+from src.services.auth_service import verify_api_key, verify_jwt
 from src.services.document_processor import DocumentProcessor
 from src.models.request_models import TextRequest, URLsRequest
 
@@ -12,7 +13,12 @@ def get_document_processor() -> DocumentProcessor:
 
 # delete documents with source label from query (/?source_label=...)
 @router.delete("/")
-async def delete_documents_with_source_label(source_label: str, processor: DocumentProcessor = Depends(get_document_processor)):
+async def delete_documents_with_source_label(authorization: Annotated[str, Header()], source_label: str, processor: DocumentProcessor = Depends(get_document_processor)):
+    """
+    Delete documents with a specific source label.
+    """
+    if not verify_jwt(authorization.split(' ')[-1]):
+        raise HTTPException(status_code=401, detail='Unauthorized: Invalid API key')
     try:
         processor.vector_store.delete_embeddings_by_source_label(source_label)
         return {}
@@ -21,10 +27,12 @@ async def delete_documents_with_source_label(source_label: str, processor: Docum
         raise HTTPException(status_code=500, detail=f"Error deleting documents: {str(e)}")
 
 @router.post("/text")
-async def process_text(request: TextRequest, processor: DocumentProcessor = Depends(get_document_processor)):
+async def process_text(authorization: Annotated[str, Header()], request: TextRequest, processor: DocumentProcessor = Depends(get_document_processor)):
     """
     Process raw text by sending it to the document processor.
     """
+    if not verify_jwt(authorization.split(' ')[-1]):
+        raise HTTPException(status_code=401, detail='Unauthorized: Invalid API key')
     try:
         processor.process_text(request.text, request.title)
         return {"message": "Text processed successfully."}
@@ -33,10 +41,12 @@ async def process_text(request: TextRequest, processor: DocumentProcessor = Depe
         raise HTTPException(status_code=500, detail=f"Error processing text: {str(e)}")
 
 @router.post("/pdf")
-async def process_pdf(title: Annotated[str, Form()], file: UploadFile = File(...), processor: DocumentProcessor = Depends(get_document_processor)):
+async def process_pdf(authorization: Annotated[str, Header()], title: Annotated[str, Form()], file: UploadFile = File(...), processor: DocumentProcessor = Depends(get_document_processor)):
     """
     Process a PDF file by sending it to the document processor.
     """
+    if not verify_jwt(authorization.split(' ')[-1]):
+        raise HTTPException(status_code=401, detail='Unauthorized: Invalid API key')
     try:
         file_bytes = await file.read()
         
@@ -47,10 +57,12 @@ async def process_pdf(title: Annotated[str, Form()], file: UploadFile = File(...
         raise HTTPException(status_code=500, detail=f"Error processing PDF: {str(e)}")
 
 @router.post("/url")
-async def process_urls(request: URLsRequest, processor: DocumentProcessor = Depends(get_document_processor)):
+async def process_urls(authorization: Annotated[str, Header()], request: URLsRequest, processor: DocumentProcessor = Depends(get_document_processor)):
     """
     Process a list of URLs by sending them to the document processor.
     """
+    if not verify_jwt(authorization.split(' ')[-1]):
+        raise HTTPException(status_code=401, detail='Unauthorized: Invalid API key')
     try:
         processor.process_urls(request.urls)
         return {"message": "URLs processed successfully."}
